+++ tests/src/utils/genSmoothRandNums.h
@@ -3,8 +3,9 @@
 
 #include "include/zfp/types.h"
 
-int
-intPow(int base, int exponent);
+// used to compute (square) array sizes
+size_t
+intPow(size_t base, int exponent);
 
 // a double pointer is passed because memory allocation
 // is taken care of within the functions
@@ -12,21 +13,21 @@ intPow(int base, int exponent);
 // generate randomly correlated integers in range:
 // [-(2^amplitudeExp - 1), 2^amplitudeExp - 1] (64 bit)
 void
-generateSmoothRandInts64(int minTotalElements, int numDims, int amplitudeExp, int64** outputArr, int* outputSideLen, int* outputTotalLen);
+generateSmoothRandInts64(size_t minTotalElements, int numDims, int amplitudeExp, int64** outputArr, size_t* outputSideLen, size_t* outputTotalLen);
 
 // generate randomly correlated integers in range:
 // [-(2^amplitudeExp - 1), 2^amplitudeExp - 1] (32 bit)
 void
-generateSmoothRandInts32(int minTotalElements, int numDims, int amplitudeExp, int32** outputArr32Ptr, int* outputSideLen, int* outputTotalLen);
+generateSmoothRandInts32(size_t minTotalElements, int numDims, int amplitudeExp, int32** outputArr32Ptr, size_t* outputSideLen, size_t* outputTotalLen);
 
 // generate randomly correlated floats in range:
 // [-(2^11), 2^11 - 2^(-12)]
 void
-generateSmoothRandFloats(int minTotalElements, int numDims, float** outputArrPtr, int* outputSideLen, int* outputTotalLen);
+generateSmoothRandFloats(size_t minTotalElements, int numDims, float** outputArrPtr, size_t* outputSideLen, size_t* outputTotalLen);
 
 // generate randomly correlated doubles in range:
 // [-(2^26), 2^26 - 2^(-26)]
 void
-generateSmoothRandDoubles(int minTotalElements, int numDims, double** outputArrPtr, int* outputSideLen, int* outputTotalLen);
+generateSmoothRandDoubles(size_t minTotalElements, int numDims, double** outputArrPtr, size_t* outputSideLen, size_t* outputTotalLen);
 
 #endif
